requirements.txt
fastapi
pillow
torchvision
torch
ultralytics
opencv-python-headless
numpy
pyyaml
motor
uvicorn
pydantic
pymongo
openai
ultralytics

langchain
langchain_community
tiktoken
langchainhub
langgraph
tavily-python
gpt4all
marker-pdf
markdown
unstructured
langchain-chroma
unstructured
markdown